+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * expose `CurlCommand` interface.
 * handle `-A/--user-agent` option, expose `userAgent?: string`
 * handle `--compressed` option, expose `compressed?: boolean`
+* handle `-k/--insecure` option, expose `insecure?: boolean`
 
 ## 0.3.2